Update game
I fixed a bug where if a player hits a enemy then the player goes back to level one and if they hit 'j' key it takes them to the level they were on. what should happen is if they hit a enemy it should restart the level they're on.

Made boxes look more like platforms.

Changed it so the game and timer only starts when a player hits either w, a, s, or d.

Made it so you can hit k to restart at any time

Added a 'best time' and milliseconds.